front-end: Add a click method for interacting with the canvas
The reason for doing it this way is to ensure that we don't have side effects on click because we forgot to remove a handler. This way, changing modes replaces the previous behavior rather than adding to it.

In the future if we find that we need separate actions for right and left click, it should be easy to adjust the click handler for that case.